[AN] refactor: 상세 화면 도메인 모델 패키지 이동 #103
