+++ android/app/src/main/java/com/onair/hearit/data/HearitMapper.kt
@@ -5,9 +5,9 @@ import com.onair.hearit.data.dto.HearitResponse
 import com.onair.hearit.data.dto.RandomHearitResponse
 import com.onair.hearit.data.dto.RecommendHearitResponse
 import com.onair.hearit.data.dto.SearchHearitResponse
-import com.onair.hearit.domain.SingleHearit
 import com.onair.hearit.domain.model.RecommendHearit
 import com.onair.hearit.domain.model.SearchedHearit
+import com.onair.hearit.domain.model.SingleHearit
 
 fun SearchHearitResponse.toDomain(): SearchedHearit =
     SearchedHearit(
