+++ android/app/src/main/java/com/onair/hearit/data/repository/MediaFileRepositoryImpl.kt
@@ -2,12 +2,12 @@ package com.onair.hearit.data.repository
 
 import RandomHearitItem
 import com.onair.hearit.data.datasource.MediaFileRemoteDataSource
-import com.onair.hearit.domain.Hearit
-import com.onair.hearit.domain.OriginalAudioUrl
-import com.onair.hearit.domain.SingleHearit
+import com.onair.hearit.domain.model.Hearit
 import com.onair.hearit.domain.model.HearitShorts
+import com.onair.hearit.domain.model.OriginalAudioUrl
 import com.onair.hearit.domain.model.ScriptLine
 import com.onair.hearit.domain.model.ShortAudioUrl
+import com.onair.hearit.domain.model.SingleHearit
 import com.onair.hearit.domain.repository.MediaFileRepository
 import com.onair.hearit.domain.toHearit
 import com.onair.hearit.domain.toHearitShorts
