+++ android/app/src/main/java/com/onair/hearit/domain/model/OriginalAudioUrl.kt
@@ -1,4 +1,4 @@
-package com.onair.hearit.domain
+package com.onair.hearit.domain.model
 
 data class OriginalAudioUrl(
     val id: Long,
