+++ android/app/src/main/java/com/onair/hearit/domain/repository/MediaFileRepository.kt
@@ -1,12 +1,12 @@
 package com.onair.hearit.domain.repository
 
 import RandomHearitItem
-import com.onair.hearit.domain.Hearit
-import com.onair.hearit.domain.OriginalAudioUrl
-import com.onair.hearit.domain.SingleHearit
+import com.onair.hearit.domain.model.Hearit
 import com.onair.hearit.domain.model.HearitShorts
+import com.onair.hearit.domain.model.OriginalAudioUrl
 import com.onair.hearit.domain.model.ScriptLine
 import com.onair.hearit.domain.model.ShortAudioUrl
+import com.onair.hearit.domain.model.SingleHearit
 
 interface MediaFileRepository {
     suspend fun getShortAudioUrl(hearitId: Long): Result<ShortAudioUrl>
