+++ android/app/src/main/java/com/onair/hearit/presentation/detail/PlayerDetailViewModel.kt
@@ -5,7 +5,7 @@ import androidx.lifecycle.MutableLiveData
 import androidx.lifecycle.ViewModel
 import androidx.lifecycle.viewModelScope
 import com.onair.hearit.R
-import com.onair.hearit.domain.Hearit
+import com.onair.hearit.domain.model.Hearit
 import com.onair.hearit.domain.usecase.GetHearitUseCase
 import com.onair.hearit.presentation.SingleLiveData
 import kotlinx.coroutines.launch
