+++ android/app/src/main/res/layout/activity_player_detail.xml
@@ -7,7 +7,7 @@
 
         <variable
             name="hearit"
-            type="com.onair.hearit.domain.Hearit" />
+            type="com.onair.hearit.domain.model.Hearit" />
     </data>
 
     <androidx.constraintlayout.widget.ConstraintLayout
